handle sigint in modules, bomb info persistence
+ Handle SIGINT when running application functions to return to main
shell. This adds a dependency on pthread (each function runs in its own
thread, that gets cancelled).

* Remove unused char*[] arguments, and add struct bomb* argument. This
makes it possible for function to prompt and save information on a bomb,
so that it only prompts the user once. (This is currently only
implemented for the battery count used by complex_wires, and you can not
reset the value (yet))

+ Moved complex_wires' input functions to input.h/c